Add getConsensusProofPeers API.
