+++ lib/plugin.js
@@ -3,16 +3,24 @@
  */
 'use strict';
 
+const blocks = require('./blocks');
 const events = require('./events');
-
-// TODO: all consensus queries that interact directly with storage collections
-// should be added here, they will extend the storage classes.
 
 module.exports = {
   type: 'ledgerStoragePlugin',
   api: {
     expandIndexes: async ({createIndexes, collections}) => {
       await createIndexes([{
+        collection: collections.blockCollection,
+        fields: {
+          'block.blockHeight': 1,
+          'block.consensusProofHash': 1,
+        },
+        options: {
+          sparse: false, unique: true, background: false,
+          name: 'block.continuity2017.blockHeight.1',
+        }
+      }, {
         collection: collections.eventCollection,
         fields: {
           'meta.continuity2017.type': 1,
@@ -65,6 +73,7 @@ module.exports = {
       }]);
     },
     storage: {
+      blocks,
       events,
     },
   }
